Task/order service/432 vo money (#437)
* Feat: money 값 객체 구현

Resolves: #432

* Feat: 장바구니 Money 객체 사용 구조로 수정

Resolves: #432

* Test: 테스트 코드 수정

Resolves: #432

* Feat: 주문 시트 로직 Money 값객체 사용 로직으로 수정

Resolves: #432

* Feat: 주문 시트 도메인 Money 객체 사용

Resolves: #432

* Feat: 주문 로직 Money 값 객체 사용으로 변경

Resolves: #432

+++ order-service/src/main/java/com/example/order_service/cart/api/dto/response/CartResponse.java
@@ -57,7 +57,7 @@ public class CartResponse {
                     .thumbnail(result.thumbnail())
                     .quantity(result.quantity())
                     .price(CartItemPrice.from(result.price()))
-                    .lineTotal(result.lineTotal())
+                    .lineTotal(result.lineTotal().longValue())
                     .options(mapToOptions(result.options()))
                     .build();
         }
@@ -80,10 +80,10 @@ public class CartResponse {
     ) {
         public static CartItemPrice from(CartResult.CartItemPrice price){
             return CartItemPrice.builder()
-                    .originalPrice(price.originalPrice())
+                    .originalPrice(price.originalPrice().longValue())
                     .discountRate(price.discountRate())
-                    .discountAmount(price.discountAmount())
-                    .discountedPrice(price.discountedPrice())
+                    .discountAmount(price.discountAmount().longValue())
+                    .discountedPrice(price.discountedPrice().longValue())
                     .build();
         }
     }
